[refactor] 매칭정보 모델 파일명도 변경 #253
